pyproject.toml
[project]
name = "webscraping-cub-sc"
version = "0.1.0"
description = ""
authors = [
    {name = "Cezar Maldini",email = "cezar.datascience@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
